+++ Package/Models/ProviderOptions/FileProviderOptions.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace GrumpyCoders.BladeState.Models.ProviderOptions;
 
-public class FileProviderOptions
+public class FileSystemProviderOptions
 {
     public string BasePath { get; set; } = Environment.GetEnvironmentVariable("tmp");
 }